Function to reconstruct individual patient data (IPD) from Kaplan-Meier (KM) survival curves based on P. Guyot et al. Enhanced secondary analysis of survival data: reconstructing the data from published Kaplan-Meier survival curves, 2012 The function returns the reconstructed individual patient data (IPD) based on digitized data from Kaplan-Meier (KM) survival curves, and a dataframe with the time points or intervals at which the K-M estimates are calculated, estimated number of patients at risk at each time point, estimated number of events (occurrences of the event) at each time point, and the estimated number of censored observations at each time point.

Function to reconstruct individual patient data (IPD) from Kaplan-Meier (KM) survival curves based on P. Guyot et al. Enhanced secondary analysis of survival data: reconstructing the data from published Kaplan-Meier survival curves, 2012 The function returns the reconstructed individual patient data (IPD) based on digitized data from Kaplan-Meier (KM) survival curves, and a dataframe with the time points or intervals at which the K-M estimates are calculated, estimated number of patients at risk at each time point, estimated number of events (occurrences of the event) at each time point, and the estimated number of censored observations at each time point.

get_ipd_guyot(time, prob, trisk, nrisk, lower, upper, tot_events = NA)

Arguments

time

A vector of the x coordinates of the survival times.

prob

A vector of the y coordinates corresponding to time, between 0 and 1. These represent survival probabilities at the times in time.

trisk

A vector of times at which patients are at risk.

nrisk

A vector of the number of patients at risk.

lower

A vector that contains the lower time limits for each interval. It defines the starting point of each interval where calculations for survival estimates and events will be made.

upper

A The upper parameter is also a vector that contains the upper time limits for each interval. It defines the end point of each interval where calculations will stop.

tot_events

The number total events (optional).

Value

A list with four data.frames, one data.frame of reconstructed IPD with two columns: (1) time: survival time, (2) event: indicator of event, where 1=event and 0=censored, a data.frame with estimated parameter values, a data.frame with survival times and probabilities, potentially modified to fix non-decreasing survival probabilities, and a data.frame with details in the itme intervals.