Function to reconstruct individual patient data (IPD) from Kaplan-Meier (KM) survival curves based on B. Rogula et al. A Method for Reconstructing Individual Patient Data From Kaplan-Meier Survival Curves That Incorporate Marked Censoring Times, 2022. The function returns the reconstructed individual patient data (IPD) based on digitized data from Kaplan-Meier (KM) survival curves. As an input a vector of digitized censoring times and will incorporate them into the IPD exactly as specified. If the KM curve is horizontal at its end, either (1) the coordinates of the end of the curve can be included in time and prob, OR (2) the x-coordinate of the end of the curve can be included in cens_t.

Function to reconstruct individual patient data (IPD) from Kaplan-Meier (KM) survival curves based on B. Rogula et al. A Method for Reconstructing Individual Patient Data From Kaplan-Meier Survival Curves That Incorporate Marked Censoring Times, 2022. The function returns the reconstructed individual patient data (IPD) based on digitized data from Kaplan-Meier (KM) survival curves. As an input a vector of digitized censoring times and will incorporate them into the IPD exactly as specified. If the KM curve is horizontal at its end, either (1) the coordinates of the end of the curve can be included in time and prob, OR (2) the x-coordinate of the end of the curve can be included in cens_t.

get_ipd_rogula(n, time, prob, cens_t = NA)

Arguments

n

The number of patients at risk at time zero.

time

A vector of the x coordinates of the bottoms of the drops in survival.

prob

A vector of the y coordinates corresponding to 't', between 0 and 1. These represent survival proportions at the times in t.

cens_t

A vector of times at which patients were censored (optional).

Value

A list with two data.frames, one with the reconstructed IPD with two columns: (1) time: survival time, (2) event: indicator of event, where 1=event and 0=censored. The other one contains survival times and probabilities, potentially modified to fix non-decreasing survival probabilities.